Tidy network iface API and printf casts
Cleans up `network_iface.h` by removing redundant `inline` from `constexpr` helpers, formatting `AddressingMode()` for clarity, and renaming the `GetCounters` output parameter to `counters` for consistency. It also fixes `printf` argument types in ARP and network params debug output by casting to unsigned types that match `%u`, reducing format/type mismatch warnings.

+++ lib-network/src/core/ipv4/arp.cpp
@@ -244,7 +244,7 @@ template <network::arp::EthSend S> static void Query(uint32_t destination_ip, vo
             network::memory::Allocator::Instance().Free(record_found->packet.p);
         }
 
-        printf("size=%u\n", size);
+        printf("size=%u\n", static_cast<unsigned>(size));
         assert(size <= network::memory::kBlockSize);
         record_found->packet.p = network::memory::Allocator::Instance().Allocate();
         assert(record_found->packet.p != nullptr);
